fixed a bug where lastValue in logEntry was string if you tried adding a value that can be added as float(not encloding float)

+++ src/main/java/frc/demacia/utils/Log/LogManager.java
@@ -118,7 +118,7 @@ public class LogManager extends SubsystemBase {
       }
     } else {
         try {
-            double d = ((Number) value).floatValue();
+            float d = ((Number) value).floatValue();
             isFloat = true;
         } catch (Exception e) {
           if (value instanceof  Boolean){
